moved logger code to lib/logger.js and uuid generator to lib/utils
