+++ listServer.js
@@ -25,59 +25,9 @@ const REVISION_VER = 3;
 // ---------------
 // Require some essential libraries and modules.
 // ---------------
-const log4js = require("log4js");
-const fs = require("fs");
 const uuid = require("uuid");
+const loggerInstance = require("./lib/logger");
 const { configuration } = require("./lib/config");
-// ---------------
-// Used to store our configuration file data.
-// ---------------
-
-var logFile = "NodeListServer.log";
-
-// ---------------
-// Logging configuration. Feel free to modify.
-// ---------------
-
-// Bugfix: If Google App Engine is running NodeLS, we cannot write to the disk, since it's readonly.
-// So we run a check to ensure we can write, if not, we just log to console.
-
-// Check if we can write?
-fs.access(__dirname, fs.constants.W_OK, function (err) {
-  if (err) {
-    console.warn("Can't write to disk; logging to file will be disabled!");
-
-    // Configure Log4js
-    log4js.configure({
-      appenders: {
-        console: { type: "stdout" },
-        default: { type: "stdout" },
-      },
-      categories: {
-        default: { appenders: ["console"], level: "debug" },
-      },
-    });
-  } else {
-    log4js.configure({
-      appenders: {
-        console: { type: "stdout" },
-        default: {
-          type: "file",
-          filename: logFile,
-          maxLogSize: 1048576,
-          backups: 3,
-          compress: true,
-        },
-      },
-      categories: {
-        default: { appenders: ["default", "console"], level: "debug" },
-      },
-    });
-  }
-});
-
-// Now we get the logger instance.
-var loggerInstance = log4js.getLogger("NodeLS");
 
 // Constant references to various modules.
 const expressServer = require("express");
@@ -118,20 +68,6 @@ expressApp.use(expressServer.urlencoded({ extended: true }));
 
 // Server memory array cache.
 var knownServers = [];
-
-// Functions used by NodeListServer
-// - Utilities
-// Generates a UUID for a newly added server.
-function generateUuid() {
-  var generatedUuid = uuid.v4();
-  var doesExist = knownServers.filter((server) => server.uuid === generatedUuid); // Used for collision check
-
-  if (doesExist.length > 0) {
-    generateUuid();
-  }
-
-  return generatedUuid;
-}
 
 // Taken from https://melvingeorge.me/blog/check-if-string-is-valid-ip-address-javascript
 function checkIfValidIP(str) {
